# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Galaxy ERP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ventas/presupuestos": {
      "post": {
        "tags": [
          "Ventas y Facturaci\u00f3n"
        ],
        "summary": "Crear Presupuesto",
        "operationId": "crear_presupuesto_api_v1_ventas_presupuestos_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresupuestoVentaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresupuestoVentaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventas/facturas": {
      "post": {
        "tags": [
          "Ventas y Facturaci\u00f3n"
        ],
        "summary": "Emitir Factura",
        "operationId": "emitir_factura_api_v1_ventas_facturas_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaVentaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacturaVentaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Ventas y Facturaci\u00f3n"
        ],
        "summary": "Listar Facturas",
        "operationId": "listar_facturas_api_v1_ventas_facturas_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaVentaResponse"
                  },
                  "title": "Response Listar Facturas Api V1 Ventas Facturas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventas/notas-credito": {
      "post": {
        "tags": [
          "Ventas y Facturaci\u00f3n"
        ],
        "summary": "Crear Nota Credito",
        "operationId": "crear_nota_credito_api_v1_ventas_notas_credito_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotaCreditoCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Crear Nota Credito Api V1 Ventas Notas Credito Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventas/facturas/{factura_id}/pdf": {
      "get": {
        "tags": [
          "Ventas y Facturaci\u00f3n"
        ],
        "summary": "Descargar Factura Pdf",
        "operationId": "descargar_factura_pdf_api_v1_ventas_facturas__factura_id__pdf_get",
        "parameters": [
          {
            "name": "factura_id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Factura Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Descargar Factura Pdf Api V1 Ventas Facturas  Factura Id  Pdf Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FacturaVentaCreate": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "warehouse_id": {
            "type": "string",
            "format": "uuid",
            "title": "Warehouse Id"
          },
          "vendedor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendedor Id"
          },
          "presupuesto_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presupuesto Id"
          },
          "fecha_emision": {
            "type": "string",
            "format": "date",
            "title": "Fecha Emision"
          },
          "moneda": {
            "type": "string",
            "maxLength": 3,
            "title": "Moneda",
            "default": "USD"
          },
          "tasa_cambio": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasa Cambio",
            "description": "Opcional: si se omite se usa tasa del d\u00eda"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/VentaItem"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "cliente_id",
          "warehouse_id",
          "fecha_emision",
          "items"
        ],
        "title": "FacturaVentaCreate"
      },
      "FacturaVentaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "warehouse_id": {
            "type": "string",
            "format": "uuid",
            "title": "Warehouse Id"
          },
          "vendedor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendedor Id"
          },
          "numero_factura": {
            "type": "string",
            "title": "Numero Factura"
          },
          "numero_control": {
            "type": "string",
            "title": "Numero Control"
          },
          "fecha_emision": {
            "type": "string",
            "format": "date",
            "title": "Fecha Emision"
          },
          "tasa_cambio": {
            "type": "string",
            "title": "Tasa Cambio"
          },
          "moneda": {
            "type": "string",
            "title": "Moneda"
          },
          "monto_exento": {
            "type": "string",
            "title": "Monto Exento"
          },
          "base_imponible": {
            "type": "string",
            "title": "Base Imponible"
          },
          "monto_iva": {
            "type": "string",
            "title": "Monto Iva"
          },
          "total_usd": {
            "type": "string",
            "title": "Total Usd"
          },
          "total_ves": {
            "type": "string",
            "title": "Total Ves"
          },
          "cliente_retiene_iva": {
            "type": "boolean",
            "title": "Cliente Retiene Iva"
          },
          "porcentaje_retencion_iva": {
            "type": "string",
            "title": "Porcentaje Retencion Iva"
          },
          "monto_retencion_iva": {
            "type": "string",
            "title": "Monto Retencion Iva"
          },
          "cliente_retiene_islr": {
            "type": "boolean",
            "title": "Cliente Retiene Islr"
          },
          "porcentaje_retencion_islr": {
            "type": "string",
            "title": "Porcentaje Retencion Islr"
          },
          "monto_retencion_islr": {
            "type": "string",
            "title": "Monto Retencion Islr"
          },
          "saldo_pendiente_usd": {
            "type": "string",
            "title": "Saldo Pendiente Usd"
          },
          "saldo_pendiente_ves": {
            "type": "string",
            "title": "Saldo Pendiente Ves"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "cliente_id",
          "warehouse_id",
          "vendedor_id",
          "numero_factura",
          "numero_control",
          "fecha_emision",
          "tasa_cambio",
          "moneda",
          "monto_exento",
          "base_imponible",
          "monto_iva",
          "total_usd",
          "total_ves",
          "cliente_retiene_iva",
          "porcentaje_retencion_iva",
          "monto_retencion_iva",
          "cliente_retiene_islr",
          "porcentaje_retencion_islr",
          "monto_retencion_islr",
          "saldo_pendiente_usd",
          "saldo_pendiente_ves",
          "estado",
          "created_at"
        ],
        "title": "FacturaVentaResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotaCreditoCreate": {
        "properties": {
          "factura_id": {
            "type": "string",
            "format": "uuid",
            "title": "Factura Id"
          },
          "fecha_emision": {
            "type": "string",
            "format": "date",
            "title": "Fecha Emision"
          },
          "motivo": {
            "type": "string",
            "minLength": 5,
            "title": "Motivo"
          },
          "monto_total_usd": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Monto Total Usd"
          },
          "monto_total_ves": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Monto Total Ves"
          }
        },
        "type": "object",
        "required": [
          "factura_id",
          "fecha_emision",
          "motivo",
          "monto_total_usd",
          "monto_total_ves"
        ],
        "title": "NotaCreditoCreate"
      },
...
      "PresupuestoVentaCreate": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "warehouse_id": {
            "type": "string",
            "format": "uuid",
            "title": "Warehouse Id"
          },
          "vendedor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendedor Id"
          },
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "vigencia_dias": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Vigencia Dias",
            "default": 15
          },
          "moneda": {
            "type": "string",
            "maxLength": 3,
            "title": "Moneda",
            "default": "USD"
          },
          "notas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notas"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/VentaItem"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "cliente_id",
          "warehouse_id",
          "fecha",
          "items"
        ],
        "title": "PresupuestoVentaCreate"
      },
      "PresupuestoVentaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "numero": {
            "type": "string",
            "title": "Numero"
          },
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "tasa_cambio": {
            "type": "string",
            "title": "Tasa Cambio"
          },
          "moneda": {
            "type": "string",
            "title": "Moneda"
          },
          "total_usd": {
            "type": "string",
            "title": "Total Usd"
          },
          "total_ves": {
            "type": "string",
            "title": "Total Ves"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "cliente_id",
          "numero",
          "fecha",
          "tasa_cambio",
          "moneda",
          "total_usd",
          "total_ves",
          "estado",
          "created_at"
        ],
        "title": "PresupuestoVentaResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VentaItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "cantidad": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Cantidad"
          },
          "precio_unitario": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Precio Unitario"
          },
          "alicuota_iva": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Alicuota Iva",
            "default": "16.00"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "cantidad",
          "precio_unitario"
        ],
        "title": "VentaItem"
      },
...
    }
  }
}